Initial pull over from the split

+++ core/src/com/projectkorra/projectkorra/ability/LightningAbility.java
@@ -39,7 +39,7 @@ public abstract class LightningAbility extends FireAbility implements SubAbility
 
 	@Override
 	public long applyModifiersCooldown(long value) {
-		return (long) GeneralMethods.applyInverseModifiers(value, getDayFactor(1.0));
+		return GeneralMethods.applyInverseModifiers(value, getDayFactor(1.0));
 	}
 
 }
